update to getRate and getExchange methods
This commit contains changes to the getRate method. It takes and walks through until it finds currency code. Then stores as a JsonObject. The object is converted to a float then returned.

Each method has a return statement for the getRate method. The second method takes and returns the differences between the two currency exchange rates.